+++ cloudx_flutter_sdk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2025-11-12
+
 ### Changed
 - **BREAKING**: iOS SDK support temporarily disabled - iOS native SDK not yet ready for production
   - CloudXCore pod dependency removed from iOS plugin